Further clean-up and better modularity.
