+++ src/cloudsc_python/cloudsc4py/physics/_stencils/__init__.py
@@ -3,3 +3,4 @@ import cloudsc4py.physics._stencils.cloudsc
 import cloudsc4py.physics._stencils.cuadjtq
 import cloudsc4py.physics._stencils.fccld
 import cloudsc4py.physics._stencils.fcttre
+import cloudsc4py.physics._stencils.helpers
